Chamge to 32bit

+++ utils.py
@@ -16,7 +16,7 @@ from time import gmtime, strftime
 from six.moves import xrange
 from PIL import Image
 
-import tensorflow as tf
+import tensorflow as tffromarray
 import tensorflow.contrib.slim as slim
 
 pp = pprint.PrettyPrinter()
@@ -91,7 +91,7 @@ def center_crop(x, crop_h, crop_w,
   h, w = x.shape[:2]
   j = int(round((h - crop_h)/2.))
   i = int(round((w - crop_w)/2.))
-  im = Image.fromarray(x[j:j+crop_h, i:i+crop_w])
+  im = Image.fromarray(np.float32(x[j:j+crop_h, i:i+crop_w]))
   return np.array(im.resize([resize_h, resize_w]), PIL.Image.BILINEAR)
 
 def transform(image, input_height, input_width, 
@@ -101,7 +101,7 @@ def transform(image, input_height, input_width,
       image, input_height, input_width, 
       resize_height, resize_width)
   else:
-    im = Image.fromarray(image[j:j+crop_h, i:i+crop_w])
+    im = Image.fromarray(np.float32(image[j:j+crop_h, i:i+crop_w]))
   return np.array(im.resize([resize_h, resize_w]), PIL.Image.BILINEAR)/127.5 - 1.
 
 def inverse_transform(images):
